Overhauls the implementation to fix a couple of race condition issues.
The current implementation ignores the ack/nack of the last byte even when
it should be checked on write operations.
Another recurrent issue is with rx fifo overflow possibly leading to push
stall and freeze the whole system. A previous commit attempted to fix that
but there's likely still cases where this may happen.
It is also unable to discriminate a Nack on the address nor data.

This implementation fixes all the previously cited issue by making sure all
operation start and end with no transmission operation "on flight".
The start/stop/restart operation are special in that they don't generate
rx bytes.